+++ components/ChartData.tsx
@@ -8,6 +8,10 @@ import gql from "graphql-tag";
 import { useQuery } from "@apollo/react-hooks";
 import TablePagination from "@material-ui/core/TablePagination";
 import numbro from "numbro";
+import ComponentLoader from './misc/ComponentLoader';
+import NotAvailable from './misc/NotAvailable'
+import ErrorMessage from './misc/ErrorMessage';
+
 import { GET_CHAIN } from './query/Chain'
 
 
@@ -40,10 +44,6 @@ const useStyles = makeStyles({
 
 });
 
-const NotAvailable = () => {
-  const classes = useStyles();
-  return <span className={classes.label}>Data not available</span>
-}
 
 const ChartData = () => {
   const classes = useStyles();
@@ -52,8 +52,8 @@ const ChartData = () => {
     pollInterval: 5000,
   });
 
-  if (loading) return <>{"Loading..."}</>
-  if (error) return <>{`Error! ${error.message}`}</>
+  if (loading) return <ComponentLoader size="small" />
+  if (error) return <ErrorMessage message={error.message} />
 
 
   return (
@@ -68,7 +68,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 $ {numbro(data.chain.tokenPrice.usd).format("0.00")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -81,7 +81,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 $ {numbro(data.chain.tokenPrice.usdMarketCap).format("0.00")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -94,7 +94,7 @@ const ChartData = () => {
               <><Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.averageBlockTime).format("0.00")}
               </Typography><span className={classes.valueSuffix}>seconds</span></>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -107,7 +107,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.txCount).format("000,000")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -120,7 +120,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.latestHeight).format("000,000")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -133,7 +133,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.walletCount).format("000,000")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
       </Grid>