+++ components/block/BlockDetails.tsx
@@ -14,6 +14,9 @@ import gql from "graphql-tag";
 import { useQuery } from "@apollo/react-hooks";
 import ContentLoader from "react-content-loader";
 import moment from "moment";
+import ComponentLoader from '../misc/ComponentLoader';
+import NotAvailable from '../misc/NotAvailable'
+import ErrorMessage from '../misc/ErrorMessage';
 import { GET_BLOCK_DETAILS } from '../query/Block'
 
 
@@ -54,21 +57,20 @@ const useStyles = makeStyles(() => {
 });
 
 const BlockDetails = () => {
-  // // const BlockDetails = (number_value : any  ) => {
-  // export default function Block(number_value: any) {
   const router = useRouter();
 
   if (!router.query.block) return <ContentLoader />;
-  //console.log(router.query.block);
-  const number = parseInt((router.query.block).toString());
+
+  //parse to String first to satisfy typecheck 
+  const number = parseInt(router.query.block.toString());
   const prevBlock: number = number - 1;
   const nextBlock: number = number + 1;
   const { loading, error, data } = useQuery(GET_BLOCK_DETAILS, {
     variables: { number },
   });
   const classes = useStyles();
-  if (loading) return null;
-  if (error) return <>{`Error! ${error.message}`}</>
+  if (loading) return <ComponentLoader />
+  if (error) return <ErrorMessage message={error.message} />
   return (
     <Card className={classes.root}>
       <CardContent>
@@ -119,7 +121,7 @@ const BlockDetails = () => {
             <Typography variant="body2" component="h2">
               {data.block && data.block.timestamp
                 ? new Date(parseInt(data.block.timestamp) * 1000).toUTCString()
-                : "Data currently not available"}{" "}
+                : <NotAvailable variant="body2" />}
               (
               {data && data.block && data.block.timestamp
                 ? moment.unix(data.block.timestamp).fromNow()
@@ -137,7 +139,7 @@ const BlockDetails = () => {
                 data.block.transactions &&
                 data.block.transactions.transactionIndex
                 ? data.block.transactions.transactionIndex.length()
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -147,7 +149,7 @@ const BlockDetails = () => {
             <Typography variant="body2" component="h2">
               {data.block && data.block.size
                 ? data.block.size
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -164,9 +166,7 @@ const BlockDetails = () => {
                   <Link href="#" color="secondary">
                     {data.block.miner.name}
                   </Link>
-                ) : (
-                  "Data currently not available"
-                )}
+                ) : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -178,7 +178,7 @@ const BlockDetails = () => {
             <Typography variant="body2" component="h2">
               {data.block && data.block.hash
                 ? data.block.hash
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -198,9 +198,7 @@ const BlockDetails = () => {
                 >
                   {data.block.parentHash}
                 </Link>
-              ) : (
-                  "Data currently not available"
-                )}
+              ) : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -212,7 +210,7 @@ const BlockDetails = () => {
             <Typography variant="body2" component="h2">
               {data.block && data.block.totalDifficulty
                 ? data.block.totalDifficulty
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -226,7 +224,7 @@ const BlockDetails = () => {
                 data.block.transactions &&
                 data.block.transactions.nonce
                 ? data.block.transactions.nonce
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -238,7 +236,7 @@ const BlockDetails = () => {
             <Typography variant="body2" component="h2">
               {data.block && data.block.gasUsed
                 ? data.block.gasUsed
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -250,7 +248,7 @@ const BlockDetails = () => {
             <Typography variant="body2" component="h2">
               {data.block && data.block.gasLimit
                 ? data.block.gasLimit
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>