+++ components/ledger/proposal/vote/Confirm.tsx
@@ -8,6 +8,7 @@ import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
 import Divider from "@material-ui/core/Divider";
 import IconButton from "@material-ui/core/IconButton";
+import NotAvailable from '../../../misc/NotAvailable'
 
 
 
@@ -124,7 +125,7 @@ const Confirm = () => {
                                     10
                 {/* {data.block && data.block.timestamp
                 ? new Date(parseInt(data.block.timestamp) * 1000).toUTCString()
-                : "Data currently not available"}{" "}
+                : <NotAvailable variant="body2" />}
               (
               {data && data.block && data.block.timestamp
                 ? moment.unix(data.block.timestamp).fromNow()
@@ -145,7 +146,7 @@ const Confirm = () => {
               data.block.transactions &&
               data.block.transactions.transactionIndex
                 ? data.block.transactions.transactionIndex.length()
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
                                 </Typography>
                                 <Divider variant="middle" className={classes.divider} />
                             </Grid>
@@ -158,7 +159,7 @@ const Confirm = () => {
                                     Proposal
                 {/* {data.block && data.block.size
                 ? data.block.size
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
                                 </Typography>
                                 <Divider variant="middle" className={classes.divider} />
                             </Grid>
@@ -171,7 +172,7 @@ const Confirm = () => {
                                     Don’t Burn Deposits for Rejected Governance Proposals Unless Vetoed
                 {/* {data.block && data.block.hash
                 ? data.block.hash
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
                                 </Typography>
                                 <Divider variant="middle" className={classes.divider} />
                             </Grid>
@@ -192,9 +193,7 @@ const Confirm = () => {
                 >
                   {data.block.parentHash}
                 </Link>
-              ) : (
-                "Data currently not available"
-              )} */}
+              ) : <NotAvailable variant="body2" />} */}
                                 </Typography>
                                 <Divider variant="middle" className={classes.divider} />
                             </Grid>