+++ components/proposal/ProposalDetails.tsx
@@ -15,6 +15,8 @@ import { useQuery } from "@apollo/react-hooks";
 import ContentLoader from "react-content-loader";
 import moment from "moment";
 import LedgerButtons from "../ledger/LedgerButtons";
+import NotAvailable from '../misc/NotAvailable'
+import ErrorMessage from '../misc/ErrorMessage';
 
 
 const useStyles = makeStyles(() => {
@@ -79,7 +81,7 @@ const ProposalDetails = () => {
   //   });
   const classes = useStyles();
   //   if (loading) return null;
-  //   if (error) return <>{`Error! ${error.message}`}</>
+  //   if (error) return <ErrorMessage message={error.message} />
   return (
     <Card className={classes.root}>
       <CardContent>
@@ -130,7 +132,7 @@ const ProposalDetails = () => {
             <Typography variant="body2" >
               {/* {data.block && data.block.timestamp
                 ? new Date(parseInt(data.block.timestamp) * 1000).toUTCString()
-                : "Data currently not available"}{" "}
+                : <NotAvailable variant="body2" />}
               (
               {data && data.block && data.block.timestamp
                 ? moment.unix(data.block.timestamp).fromNow()
@@ -148,7 +150,7 @@ const ProposalDetails = () => {
               data.block.transactions &&
               data.block.transactions.transactionIndex
                 ? data.block.transactions.transactionIndex.length()
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -158,7 +160,7 @@ const ProposalDetails = () => {
             <Typography variant="body2" >
               {/* {data.block && data.block.size
                 ? data.block.size
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -175,9 +177,7 @@ const ProposalDetails = () => {
                 <Link href="#" color="secondary">
                   {data.block.miner.name}
                 </Link>
-              ) : (
-                "Data currently not available"
-              )} */}
+              ) : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -189,7 +189,7 @@ const ProposalDetails = () => {
             <Typography variant="body2"  >
               {/* {data.block && data.block.hash
                 ? data.block.hash
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -209,9 +209,7 @@ const ProposalDetails = () => {
                 >
                   {data.block.parentHash}
                 </Link>
-              ) : (
-                "Data currently not available"
-              )} */}
+              ) : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -223,7 +221,7 @@ const ProposalDetails = () => {
             <Typography variant="body2"  >
               {/* {data.block && data.block.totalDifficulty
                 ? data.block.totalDifficulty
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -237,7 +235,7 @@ const ProposalDetails = () => {
               data.block.transactions &&
               data.block.transactions.nonce
                 ? data.block.transactions.nonce
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -249,7 +247,7 @@ const ProposalDetails = () => {
             <Typography variant="body2"  >
               {/* {data.block && data.block.gasUsed
                 ? data.block.gasUsed
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -261,7 +259,7 @@ const ProposalDetails = () => {
             <Typography variant="body2"  >
               {/* {data.block && data.block.gasLimit
                 ? data.block.gasLimit
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -273,7 +271,7 @@ const ProposalDetails = () => {
             <Typography variant="body2"  >
               {/* {data.block && data.block.gasLimit
                 ? data.block.gasLimit
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>