+++ components/transaction/TransactionDetails.tsx
@@ -21,6 +21,9 @@ import moment from "moment";
 import Chip from "@material-ui/core/Chip";
 import Snackbar from "@material-ui/core/Snackbar";
 import MuiAlert, { AlertProps } from "@material-ui/lab/Alert";
+import ComponentLoader from '../misc/ComponentLoader';
+import NotAvailable from '../misc/NotAvailable'
+import ErrorMessage from '../misc/ErrorMessage';
 import { GET_TX_DETAILS } from '../query/Transaction'
 
 
@@ -113,11 +116,12 @@ const TransactionDetails = (props: any) => {
 
   const handleClick = () => {
     return navigator.clipboard
-      .writeText(document.getElementById("raw-input-form").value)
-      .then(() => setOpen(true))
-      .catch((err) => {
-        console.log("Something went wrong", err);
-      });
+      .writeText(document.getElementById("raw-input-form").value) != undefined ? navigator.clipboard
+        .writeText(document.getElementById("raw-input-form").value)
+        .then(() => setOpen(true))
+        .catch((err) => {
+          console.log("Something went wrong", err);
+        }) : ''
   };
 
   const handleClose = (event?: React.SyntheticEvent, reason?: string) => {
@@ -145,8 +149,8 @@ const TransactionDetails = (props: any) => {
   const handleClickUTF8 = (props: any) => {
     return (document.getElementById("raw-input-form").value = inputValue);
   };
-  if (loading) return null;
-  if (error) return <>{`Error! ${error.message}`}</>
+  if (loading) return <ComponentLoader />
+  if (error) return <ErrorMessage message={error.message} />
   return (
     <Card className={classes.root}>
       <CardContent>
@@ -164,7 +168,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.hash
                 ? data.transaction.hash
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -178,7 +182,7 @@ const TransactionDetails = (props: any) => {
                 ? new Date(
                   parseInt(data.transaction.timestamp) * 1000
                 ).toUTCString()
-                : "Data currently not available"}{" "}
+                : <NotAvailable variant="body2" />}
               (
               {data && data.transaction && data.transaction.timestamp
                 ? moment.unix(data.transaction.timestamp).fromNow()
@@ -231,9 +235,7 @@ const TransactionDetails = (props: any) => {
                 >
                   {data.transaction.from.address}
                 </Link>
-              ) : (
-                  "Data currently not available"
-                )}
+              ) : <NotAvailable variant="body2" />}
             </Typography>
 
             <Divider variant="middle" className={classes.divider} />
@@ -252,9 +254,7 @@ const TransactionDetails = (props: any) => {
                 >
                   {data.transaction.to.address}
                 </Link>
-              ) : (
-                  "Data currently not available"
-                )}
+              ) : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -266,7 +266,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.value
                 ? data.transaction.value
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -282,7 +282,7 @@ const TransactionDetails = (props: any) => {
               >
                 {data.transaction && data.transaction.blockNumber
                   ? data.transaction.blockNumber
-                  : "Data currently not available"}
+                  : <NotAvailable variant="body2" />}
               </Link>
             </Typography>
             <Divider variant="middle" className={classes.divider} />
@@ -305,7 +305,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.nonce
                 ? data.transaction.nonce
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -317,7 +317,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.feeCurrency
                 ? data.transaction.feeCurrency + "cGLD"
-                : "Data currently not available"}{" "}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -329,7 +329,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.gatewayFeeRecipient
                 ? data.transaction.gatewayFeeRecipient
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -341,7 +341,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.gatewayFee
                 ? data.transaction.gatewayFee
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -414,7 +414,7 @@ const TransactionDetails = (props: any) => {
             <Typography variant="body2" component="h2">
               {data.transaction && data.transaction.gas
                 ? data.transaction.gas
-                : "Data currently not available"}
+                : <NotAvailable variant="body2" />}
             </Typography>
             <Divider variant="middle" className={classes.divider} />
           </Grid>